+++ packages/api/src/procedures/authed.ts
@@ -1,0 +1,6 @@
+import { withAuthed } from "../middlewares/withAuthed";
+import type { APIClientType } from "../utils";
+import { publicProcedure } from "./public";
+
+export const createAuthedProcedure = (apiClientType: APIClientType) =>
+  publicProcedure.use(withAuthed(apiClientType));