+++ packages/blob-propagator/src/BlobPropagator.ts
@@ -52,12 +52,13 @@ export type BlobPropagatorConfig = {
 
 export const STORAGE_WORKER_PROCESSORS: Record<
   BlobStorageName,
-  BlobPropagationWorkerProcessor
+  BlobPropagationWorkerProcessor | undefined
 > = {
   GOOGLE: gcsProcessor,
   SWARM: swarmProcessor,
   POSTGRES: postgresProcessor,
   FILE_SYSTEM: fileSystemProcessor,
+  WEAVEVM: undefined,
 };
 
 export class BlobPropagator {
@@ -81,25 +82,47 @@ export class BlobPropagator {
       ...workerOptions,
     };
 
-    const availableStorageNames = blobStorageManager
-      .getAllStorages()
-      .map((s) => s.name)
-      .filter((name) => name !== tmpBlobStorage);
     const temporaryBlobStorage = blobStorageManager.getStorage(tmpBlobStorage);
-
-    if (!availableStorageNames) {
-      throw new BlobPropagatorCreationError("No blob storages available");
-    }
 
     if (!temporaryBlobStorage) {
       throw new BlobPropagatorCreationError("Temporary blob storage not found");
     }
 
+    const availableStorageNames = blobStorageManager
+      .getAllStorages()
+      .map((s) => s.name)
+      .filter((name) => name !== tmpBlobStorage)
+      .filter((name) => {
+        const hasWorkerProcessor = !!STORAGE_WORKER_PROCESSORS[name];
+
+        if (!hasWorkerProcessor) {
+          logger.warn(
+            `Worker processor not defined for storage "${name}"; skipping`
+          );
+        }
+
+        return hasWorkerProcessor;
+      });
+
+    if (!availableStorageNames.length) {
+      throw new BlobPropagatorCreationError(
+        "None of the available storages have worker processors defined"
+      );
+    }
+
     this.storageWorkers = availableStorageNames.map(
       (storageName: BlobStorageName) => {
+        const workerProcessor = STORAGE_WORKER_PROCESSORS[storageName];
+
+        if (!workerProcessor) {
+          throw new BlobPropagatorCreationError(
+            `Worker processor not defined for storage "${storageName}"`
+          );
+        }
+
         return this.#createWorker(
           STORAGE_WORKER_NAMES[storageName],
-          STORAGE_WORKER_PROCESSORS[storageName]({
+          workerProcessor({
             prisma,
             blobStorageManager,
           }),